requirements.txt
Pillow==11.2.1
pyautogui==0.9.54
pywin32==308
Requests==2.32.3
torch==2.0.1+cu118
transformers==4.49.0
